Integrate preshared keys into key derivation.
The general idea is even if the asymmetric key
security is compromised (Quantum computers or otherwise),
The attacker must have had access to one of our keys to
access the pre shared portion. If they only had access to the
stored data, but not our pre shared sub keys, they still will
not be able to derive the decryption key.