Retry on fs smear when performing 'put' on a directory.
We force a send checkpoint and then retry a limited number of times
if we detect an error that indicates the filesystem was changed underneath
us in such a way our backup would become corrupt.

The main use of this feature is user ergonomics, a user can snapshot an in use
filesystem and stand a chance of a fairly consistent snapshot.
This is not recommended, but probably what naive users will do.