Remove sqlite3 from the repository format.
This change replaces sqlite3 transactions with a new module called
fstx. Fstx lets you do atomic transactions across groups of files by
writing a rollback journal to disk before making a group of changes,
on crash the resuming process is able to rollback all the changes.

This commit also changes the locking semantics such that GC and write
operations are mutually exclusive and must wait for eachother to complete.