Add checksums to client side caches.
These caches are vulnerable to corruption if the client machine has
faulty hardware, in odrder to mitigate this risk we now enable checksums
on all sqlite3 pages.

To add these checksums to our sqlite3 based caches we have embedded the
https://www.sqlite.org/cksumvfs.html extension.

Because cargo offers no way to get the sqlite3 header file bundled
by rusqlite, we include a copy, but verify the versions match with
an assertion.